Fixing warnings

+++ Fall_2014_iOS7_ObjectiveC/Matchismo/Matchismo/CardGameViewController.m
@@ -44,7 +44,7 @@
 #pragma Card Buttons
 - (IBAction)touchCardButton:(UIButton *)sender {
     
-    int chosenButtonIndex = [self.cardButtons indexOfObject:sender];
+    NSUInteger chosenButtonIndex = [self.cardButtons indexOfObject:sender];
     [self.game chooseCardAtIndex:chosenButtonIndex];
     [self updateUI];
 }
@@ -52,7 +52,7 @@
 - (void)updateUI
 {
     for (UIButton *cardButton in self.cardButtons) {
-        int cardButtonIndex = [self.cardButtons indexOfObject:cardButton];
+        NSUInteger cardButtonIndex = [self.cardButtons indexOfObject:cardButton];
         Card *card = [self.game cardAtIndex:cardButtonIndex];
         [cardButton setTitle:[self titleForCard:card]
                     forState:UIControlStateNormal];
@@ -60,7 +60,7 @@
                               forState:UIControlStateNormal];
         //disable matched cards
         cardButton.enabled = !card.isMatched;
-        self.scoreLabel.text = [NSString stringWithFormat:@"Score: %d", self.game.score];
+        self.scoreLabel.text = [NSString stringWithFormat:@"Score: %d", (int) self.game.score];
     }
 }
